Extract dummy news data from MainActivity to new creat NewsService class
1. Create NewsService class
2. Extract the dummy news data to NewsService class and can be accessed by a custom get function.
